microservices/quality-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pandas==2.1.3
numpy==1.26.2
pyyaml==6.0.1
python-multipart==0.0.6

# Database dependencies
asyncpg==0.29.0
redis==5.0.1
python-dotenv==1.0.0
